requirements.txt
beautifulsoup4==4.8.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
ez-setup==0.9
google==2.0.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.2.0
keyring==19.3.0
more-itertools==8.0.2
mysql-connector==2.2.9
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.0
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
soupsieve==1.9.5
tabulate==0.8.6
uritemplate==3.0.0
urllib3==1.25.6
yagmail==0.11.220
zipp==0.6.0
